UPdated image and cache commands
